@@ -1,11 +1,10 @@
 // Rested Core
-// Version: 0.7.0-alpha
+// Version: 0.7-alpha
 // www.restedwf.com
 // © 2023 Thomas Sebastian Berge
 
 export 'src/request.dart';
 export 'src/servercore.dart';
 export 'src/requesthandler.dart';
-export 'src/filecollection.dart';
 export 'src/resource.dart';
 export 'src/globals.dart';
